% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{as.data.frame.symbolic_histogram}
\alias{as.data.frame.symbolic_histogram}
\title{a data.frame}
\usage{
\method{as.data.frame}{symbolic_histogram}(x, ...)
}
\arguments{
\item{x}{.....}

\item{...}{...}
}
\description{
a data.frame
}
