% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{as.data.frame.symbolic_interval}
\alias{as.data.frame.symbolic_interval}
\title{convertir a data.frame}
\usage{
\method{as.data.frame}{symbolic_interval}(x, ...)
}
\arguments{
\item{x}{a symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}
}
\description{
convertir a data.frame
}
