% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{as.data.frame.symbolic_modal}
\alias{as.data.frame.symbolic_modal}
\title{Extract values}
\usage{
\method{as.data.frame}{symbolic_modal}(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Extract values
}
