% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.quantils.RSDA.r
\name{calculate.quantils.RSDA}
\alias{calculate.quantils.RSDA}
\title{quantiles.RSDA}
\usage{
calculate.quantils.RSDA(histogram.RSDA, num.quantils)
}
\arguments{
\item{histogram.RSDA}{A histogram}

\item{num.quantils}{Number of quantiles}
}
\value{
Quantiles of a Histogram
}
\description{
quantiles.RSDA
}
