% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{$.symbolic_histogram}
\alias{$.symbolic_histogram}
\title{$ operator for histograms}
\usage{
\method{$}{symbolic_histogram}(x, name)
}
\arguments{
\item{x}{.....}

\item{name}{...}
}
\description{
$ operator for histograms
}
