% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{$.symbolic_modal}
\alias{$.symbolic_modal}
\title{$ operator for modals}
\usage{
\method{$}{symbolic_modal}(x, name = c("cats", "props", "counts"))
}
\arguments{
\item{x}{.....}

\item{name}{...}
}
\description{
$ operator for modals
}
