% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{$.symbolic_set}
\alias{$.symbolic_set}
\title{$ operator for set}
\usage{
\method{$}{symbolic_set}(x, name = c("levels", "values"))
}
\arguments{
\item{x}{.....}

\item{name}{...}
}
\description{
$ operator for set
}
