% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_mcfa.R
\name{cfa.minmax}
\alias{cfa.minmax}
\title{cfa.minmax}
\usage{
cfa.minmax(
  sym.data,
  TFilas,
  TFilasMin,
  TFilasMax,
  TColumnas,
  TColumnasMin,
  TColumnasMax,
  Total,
  VP,
  VPzz
)
}
\description{
cfa.minmax
}
\keyword{internal}
