% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.scatterplot.gg.R
\name{cfa.scatterplot.gg}
\alias{cfa.scatterplot.gg}
\title{Plot Interval Scatterplot}
\usage{
cfa.scatterplot.gg(x, y, sym.data, pos.var)
}
\arguments{
\item{x}{value}

\item{y}{value}

\item{sym.data}{value}

\item{pos.var}{value}
}
\description{
Plot Interval Scatterplot
}
