% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_df.R
\name{check_quo_duplicated_names}
\alias{check_quo_duplicated_names}
\title{Check duplicated names in a quo}
\usage{
check_quo_duplicated_names(x)
}
\description{
Check duplicated names in a quo
}
\keyword{internal}
