% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{cont.distance}
\alias{cont.distance}
\title{Distance between continuous variables}
\usage{
cont.distance(x, y, q = 2)
}
\arguments{
\item{x}{A symbilic data table 1x1}

\item{y}{A symbilic data table 1x1}
}
\description{
Distance between continuous variables
}
\keyword{internal}
