% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.to.sym.inteval.j.r
\name{data.fame.to.sym.inteval}
\alias{data.fame.to.sym.inteval}
\title{data.fame.to.sym.inteval}
\usage{
data.fame.to.sym.inteval(data.df)
}
\description{
data.fame.to.sym.inteval
}
\keyword{internal}
