% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_mcfa.R
\name{data.frame.to.RSDA.inteval.table}
\alias{data.frame.to.RSDA.inteval.table}
\title{data.frame.to.RSDA.inteval.table}
\usage{
data.frame.to.RSDA.inteval.table(data.df)
}
\description{
data.frame.to.RSDA.inteval.table
}
\keyword{internal}
