% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.modal.R
\name{dist.modal}
\alias{dist.modal}
\title{Distance matrix for symbolic data}
\usage{
\method{dist}{modal}(sym.data)
}
\arguments{
\item{sym.data}{Symbolic data table}
}
\value{
return an object of class 'dist
}
\description{
Distance matrix for symbolic data
}
