% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_mcfa1}
\alias{ex_mcfa1}
\title{ex_mcfa1}
\format{An object of class \code{data.frame} with 130 rows and 5 columns.}
\usage{
ex_mcfa1
}
\description{
example for the sym.mcfa function.
}
\keyword{datasets}
