% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_mcfa2}
\alias{ex_mcfa2}
\title{ex_mcfa2}
\format{An object of class \code{data.frame} with 130 rows and 7 columns.}
\usage{
ex_mcfa2
}
\description{
example for the sym.mcfa function.
}
\keyword{datasets}
