% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3}
\alias{example3}
\title{Data Example 3}
\format{$C   F1 $I F2 F2 $H F3  M1  M2  M3 $C   F4 $S F5 E1 E2 E3 E4 $I     F6    F6 $I F7 F7 \cr
Case1 $C  2.8 $I  1  2 $H  3 0.1 0.7 0.2 $C  6.0 $S  4  e  g  k  i $I   0.00 90.00 $I  9 24 \cr
Case2 $C  1.4 $I  3  9 $H  3 0.6 0.3 0.1 $C  8.0 $S  4  a  b  c  d $I -90.00 98.00 $I -9  9 \cr
Case3 $C  3.2 $I -1  4 $H  3 0.2 0.2 0.6 $C -7.0 $S  4  2  1  b  c $I  65.00 90.00 $I 65 70 \cr
Case4 $C -2.1 $I  0  2 $H  3 0.9 0.0 0.1 $C  0.0 $S  4  3  4  c  a $I  45.00 89.00 $I 25 67 \cr
Case5 $C -3.0 $I -4 -2 $H  3 0.6 0.0 0.4 $C -9.5 $S  4  e  i  g  k $I  20.00 40.00 $I  9 40 \cr
Case6 $C  0.1 $I 10 21 $H  3 0.0 0.7 0.3 $C -1.0 $S  4  f  t  c  a $I   5.00  8.00 $I  5  8 \cr
Case7 $C  9.0 $I  4 21 $H  3 0.2 0.2 0.6 $C  0.5 $S  4  e  q  r  z $I   3.14  6.76 $I  4  6 \cr}
\usage{
data(example3)
}
\description{
This a symbolic data table with variables of continuos, interval,
histogram and set types.
}
\examples{
data(example3)
display.sym.table(example3)
}
\keyword{datasets}
