\name{facedata}
\alias{facedata}
\docType{data}
\title{
Face Data Example
}
\description{
Symbolic data matrix with all the variables of interval type
}
\usage{data("facedata")}
\format{
$I;AD;AD;$I;BC;BC;.........

HUS1;$I;168.86;172.84;$I;58.55;63.39;.........

HUS2;$I;169.85;175.03;$I;60.21;64.38;.........

HUS3;$I;168.76;175.15;$I;61.4;63.51;.........

INC1;$I;155.26;160.45;$I;53.15;60.21;.........

INC2;$I;156.26;161.31;$I;51.09;60.07;.........

INC3;$I;154.47;160.31;$I;55.08;59.03;.........

ISA1;$I;164;168;$I;55.01;60.03;.........

ISA2;$I;163;170;$I;54.04;59;.........

ISA3;$I;164.01;169.01;$I;55;59.01;.........

JPL1;$I;167.11;171.19;$I;61.03;65.01;.........

JPL2;$I;169.14;173.18;$I;60.07;65.07;.........

JPL3;$I;169.03;170.11;$I;59.01;65.01;.........

KHA1;$I;149.34;155.54;$I;54.15;59.14;.........

KHA2;$I;149.34;155.32;$I;52.04;58.22;.........

KHA3;$I;150.33;157.26;$I;52.09;60.21;.........

LOT1;$I;152.64;157.62;$I;51.35;56.22;.........

LOT2;$I;154.64;157.62;$I;52.24;56.32;.........

LOT3;$I;154.83;157.81;$I;50.36;55.23;.........

PHI1;$I;163.08;167.07;$I;66.03;68.07;.........

PHI2;$I;164;168.03;$I;65.03;68.12;.........

PHI3;$I;161.01;167;$I;64.07;69.01;.........

ROM1;$I;167.15;171.24;$I;64.07;68.07;.........

ROM2;$I;168.15;172.14;$I;63.13;68.07;.........

ROM3;$I;167.11;171.19;$I;63.13;68.03;.........
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\examples{
data(facedata) 
res.vertex.ps<-sym.interval.pc(facedata,"vertex",150,FALSE,FALSE,TRUE)
sym.scatterplot(sym.var(res.vertex.ps$sym.prin.curve,1),sym.var(res.vertex.ps$sym.prin.curve,2),
                labels=TRUE,col='red',main='PSC Face Data')
}
\keyword{datasets}
