% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{format.symbolic_interval}
\alias{format.symbolic_interval}
\title{Symbolic interval conversion functions to and from Character}
\usage{
\method{format}{symbolic_interval}(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Symbolic interval conversion functions to and from Character
}
