% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{format.symbolic_set}
\alias{format.symbolic_set}
\title{Symbolic set conversion functions to and from Character}
\usage{
\method{format}{symbolic_set}(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Symbolic set conversion functions to and from Character
}
