% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.columns.interval.r
\name{generate.columns.interval}
\alias{generate.columns.interval}
\title{generate.columns.interval}
\usage{
generate.columns.interval(data)
}
\description{
generate.columns.interval
}
\keyword{internal}
