% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.columns.set.r
\name{generate.columns.set}
\alias{generate.columns.set}
\title{generate.columns.set}
\usage{
generate.columns.set(data)
}
\description{
generate.columns.set
}
\author{
Jorge Arce Garro
}
\keyword{internal}
