% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.limits.PCA.indivduals.r
\name{get.limits.PCA.indivduals}
\alias{get.limits.PCA.indivduals}
\title{Projections onto PCA}
\usage{
get.limits.PCA.indivduals(
  sym.data,
  matrix.stan,
  min.stan,
  max.stan,
  svd,
  nn,
  mm
)
}
\arguments{
\item{sym.data}{An interval matrix}

\item{matrix.stan}{A standardized matrix}

\item{min.stan}{A matrix of minimum values standardized for each interval}

\item{max.stan}{A matrix of maximum values standardized for each interval}

\item{svd}{An eigen vectors matrix}

\item{nn}{Number of concepts}

\item{mm}{Number of variables}
}
\value{
Concept Projections onto the principal components
}
\description{
Calculate the interval projection onto the principal components
}
