% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sym.interval.limits.r
\name{get.sym.interval.limits}
\alias{get.sym.interval.limits}
\title{Interval Matrix associated to a Histogram Matrix}
\usage{
get.sym.interval.limits(x)
}
\arguments{
\item{x}{A Histogram matrix}
}
\value{
An Interval Matrix
}
\description{
Interval Matrix associated to a Histogram Matrix
}
\author{
Jorge Arce Garro
}
\keyword{internal}
