% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{get_cats}
\alias{get_cats}
\title{Extract categories}
\usage{
get_cats(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Extract categories
}
