% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_modal.R
\name{get_counts}
\alias{get_counts}
\title{Extract counts}
\usage{
get_counts(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Extract counts
}
