% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_modal.R
\name{get_data}
\alias{get_data}
\title{Extract values2}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Extract values2
}
