% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{get_props}
\alias{get_props}
\title{Extract prop}
\usage{
get_props(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Extract prop
}
