% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{gplot}
\alias{gplot}
\title{Symbolic object plot with ggplot2}
\usage{
gplot(x, ...)
}
\arguments{
\item{x}{a symbolic histogram}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Symbolic object plot with ggplot2
}
