% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{gplot.symbolic_histogram}
\alias{gplot.symbolic_histogram}
\title{Create a complete ggplot for symbolic histograms}
\usage{
\method{gplot}{symbolic_histogram}(x)
}
\arguments{
\item{x}{A symbolic histogram}

\item{probability}{Logical; if TRUE, the histogram graphic is a representation of frequencies}

\item{breaks}{A vector giving the breakpoints between histogram cells}

\item{...}{Further arguments to be passed from or to other methods.}
}
\value{
a ggplot object
}
\description{
Create a complete ggplot for symbolic histograms
}
\examples{
h <- sym_histogram(iris$Sepal.Length)
gplot(h)

}
