% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hardwoodBrito}
\alias{hardwoodBrito}
\title{Hard Wood Data Example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{symbolic_tbl}, \code{symbolic_tbl}, \code{symbolic_tbl}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.
}
\usage{
data('hardwoodBrito')
}
\description{
Symbolic Histogram matrix.
}
\examples{
\dontrun{
data(hardwoodBrito)
hardwoodBrito
}
}
\references{
Brito P. and Dias S. (2022).
Analysis of Distributional Data. CRC Press, United States of America.
}
\keyword{datasets}
