% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{hist.distance}
\alias{hist.distance}
\title{Distance between histogram variables}
\usage{
\method{hist}{distance}(x, y, q = 2)
}
\arguments{
\item{x}{A symbilic data table 1x1}

\item{y}{A symbilic data table 1x1}
}
\description{
Distance between histogram variables
}
\keyword{internal}
