% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{histogram}
\alias{histogram}
\title{Create an symbolic_histogram type object}
\usage{
histogram(x = double())
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic histogram
}
\description{
Create an symbolic_histogram type object
}
\examples{
histogram(iris$Sepal.Length)
}
