% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_prost_train}
\alias{int_prost_train}
\title{Linear regression model data example.}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 67 rows and 9 columns.
}
\usage{
data(int_prost_train)
}
\description{
Linear regression model interval-valued data example.
}
\references{
HASTIE, T., TIBSHIRANI, R. and FRIEDMAN, J. (2008). The Elements of Statistical Learning:
Data Mining, Inference and Prediction. New York: Springer.
}
\keyword{datasets}
