% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.interval.r
\name{intersection.interval}
\alias{intersection.interval}
\title{generate.columns.interval}
\usage{
intersection.interval(Interval.1, Interval.2)
}
\description{
generate.columns.interval
}
\author{
Jorge Arce Garro
}
\keyword{internal}
