% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{interval.distance}
\alias{interval.distance}
\title{Distance between interval variables}
\usage{
interval.distance(x, y, q = 2)
}
\arguments{
\item{x}{A symbilic data table 1x1}

\item{y}{A symbilic data table 1x1}
}
\description{
Distance between interval variables
}
\keyword{internal}
