% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_histogram_plot.R
\name{interval.histogram.plot}
\alias{interval.histogram.plot}
\title{Histogram plot for an interval variable}
\usage{
interval.histogram.plot(x, n.bins, ...)
}
\arguments{
\item{x}{An symbolic data table.}

\item{n.bins}{Numbers of breaks of the histogram.}

\item{...}{Arguments to be passed to the barplot method.}
}
\value{
A list with componets : frequency and histogram
}
\description{
Histogram plot for an interval variable
}
\examples{
data(oils)
res <- interval.histogram.plot(x = oils[, 3], n.bins = 3)
res
}
