% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.large.r
\name{interval.large}
\alias{interval.large}
\title{Calculate the large of each interval}
\usage{
interval.large(x)
}
\arguments{
\item{x}{An interval matrix}
}
\value{
A matrix with the large of each interval.
}
\description{
Calculate the large of each interval
}
\examples{
\dontrun{
data(oils)
interval.large(oils)
}
}
