% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.length.r
\name{interval.length}
\alias{interval.length}
\title{Lenght for interval}
\usage{
interval.length(x)
}
\arguments{
\item{x}{An interval matrix}
}
\value{
A matrix with the length of each interval.
}
\description{
Calculate the large of each interval
}
\examples{
\dontrun{
data(oils)
interval.length(oils)
}
}
