% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{is.sym.histogram}
\alias{is.sym.histogram}
\title{Symbolic histogram}
\usage{
is.sym.histogram(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_histogram and FALSE otherwise.
}
\description{
Symbolic histogram
}
\examples{
x <- sym.histogram(iris$Sepal.Length)
is.sym.histogram(x)
}
