% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{is.sym.interval}
\alias{is.sym.interval}
\title{Symbolic interval}
\usage{
is.sym.interval(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_vector and FALSE otherwise.
}
\description{
Symbolic interval
}
\examples{
x <- sym.interval(1:10)
is.sym.interval(x)
is.sym.interval("d")
}
