% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{is.sym.modal}
\alias{is.sym.modal}
\title{Symbolic modal}
\usage{
is.sym.modal(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_modal and FALSE otherwise.
}
\description{
Symbolic modal
}
\examples{
x <- sym.modal(factor(c("a", "b", "b", "l")))
is.sym.modal(x)
}
