% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{is.sym.set}
\alias{is.sym.set}
\title{Symbolic set}
\usage{
is.sym.set(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_set and FALSE otherwise.
}
\description{
Symbolic set
}
\examples{
x <- sym.set(factor(c("a", "b", "b", "l")))
is.sym.set(x)
}
