% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{is_histogram}
\alias{is_histogram}
\title{Symbolic histogram}
\usage{
is_histogram(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_histogram and FALSE otherwise.
}
\description{
Symbolic histogram
}
\examples{
x <- histogram(iris$Sepal.Length)
is_histogram(x)
}
