% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{is_sym_interval}
\alias{is_sym_interval}
\title{Symbolic interval}
\usage{
is_sym_interval(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_vector and FALSE otherwise.
}
\description{
Symbolic interval
}
\examples{
x <- sym_interval(1:10)
is_sym_interval(x)
is_sym_interval("d")
}
