% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_modal.R
\name{is_sym_modal}
\alias{is_sym_modal}
\title{Symbolic modal}
\usage{
is_sym_modal(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_modal and FALSE otherwise.
}
\description{
Symbolic modal
}
\examples{
x <- modal(c("a","b","b","l"))
is_modal(x)
}
