% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_set.R
\name{is_sym_set}
\alias{is_sym_set}
\title{Symbolic set}
\usage{
is_sym_set(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
returns TRUE if its argument's value is a symbolic_set and FALSE otherwise.
}
\description{
Symbolic set
}
\examples{
x <- sym_set(c("a","b","b","l"))
is_sym_set(x)
}
