% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.histogram.disjoint.pca.variable.r
\name{limits.histogram.disjoint.pca.variable}
\alias{limits.histogram.disjoint.pca.variable}
\title{limits.histogram.disjoint.pca.variable}
\usage{
limits.histogram.disjoint.pca.variable(df.histogram, BIN.Matrix)
}
\arguments{
\item{df.histogram}{Bin's Projections onto principal components}

\item{BIN.Matrix}{Number of Bins for each histogram projections}
}
\value{
Histogram Projection onto principal components
}
\description{
limits.histogram.disjoint.pca.variable
}
\author{
Jorge Arce Garro
}
\keyword{internal}
