% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.histogram.pca.r
\name{limits.histogram.pca}
\alias{limits.histogram.pca}
\title{limits.histogram.pca}
\usage{
limits.histogram.pca(sym.hist.matrix, pca.sym.interval)
}
\arguments{
\item{sym.hist.matrix}{A Histogram Matrix}

\item{pca.sym.interval}{PCA result of Interval's PCA}
}
\value{
Bin's Projections onto principal components
}
\description{
limits.histogram.pca
}
\keyword{internal}
