\name{lynne1}
\alias{lynne1}
\docType{data}
\title{
Symbolic interval data example
}
\description{
Symbolic data matrix with all the variables of interval type
}
\usage{data(lynne1)}
\format{
   $I Pulse Rate Pulse Rate $I Systolic Pressure Systolic Pressure $I Diastolic Pressure
   
1  $I         44         68 $I                90               110 $I                 50

2  $I         60         72 $I                90               130 $I                 70

3  $I         56         90 $I               140               180 $I                 90

4  $I         70        112 $I               110               142 $I                 80

5  $I         54         72 $I                90               100 $I                 50

6  $I         70        100 $I               134               142 $I                 80

7  $I         72        100 $I               130               160 $I                 76

8  $I         76         98 $I               110               190 $I                 70

9  $I         86         96 $I               138               180 $I                 90

10 $I         86        100 $I               110               150 $I                 28

   Diastolic Pressure
   
1                  70

2                  90

3                 100

4                 108

5                  70

6                 110

7                  90

8                 110

9                 110

10                100	
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\examples{
data(lynne1)
display.sym.table(lynne1)
}
\keyword{datasets}
