\name{lynne2}
\alias{lynne2}
\docType{data}
\title{
Symbolic interval data example
}
\description{
Symbolic data matrix with all the variables of interval type
}
\usage{data(lynne2)}
\format{
   $I Pulse Rate Pulse Rate $I Systolic Pressure Systolic Pressure $I Diastolic Pressure
   
1  $I         44         68 $I                90               110 $I                 50

2  $I         60         72 $I                90               130 $I                 70

3  $I         56         90 $I               140               180 $I                 90

4  $I         70        112 $I               110               142 $I                 80

5  $I         54         72 $I                90               100 $I                 50

6  $I         70        100 $I               134               142 $I                 80

7  $I         72        100 $I               130               160 $I                 76

8  $I         76         98 $I               110               190 $I                 70

9  $I         86         96 $I               138               180 $I                 90

10 $I         86        100 $I               110               150 $I                 78

11 $I         53         55 $I               160               190 $I                205

12 $I         50         55 $I               180               200 $I                110

13 $I         73         81 $I               125               138 $I                 78

14 $I         60         75 $I               175               194 $I                 90

15 $I         42         52 $I               105               115 $I                 70

   Diastolic Pressure
   
1                  70

2                  90

3                 100

4                 108

5                  70

6                 110

7                  90

8                 110

9                 110

10                100

11                219

12                125

13                 99

14                100

15                 82
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\examples{
data(lynne2)
display.sym.table(lynne2)
}
\keyword{datasets}
