% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{map_symbolic_tbl}
\alias{map_symbolic_tbl}
\title{map function over symbolic table}
\usage{
map_symbolic_tbl(.x = NULL, .f = NULL, ...)
}
\description{
map function over symbolic table
}
\keyword{internal}
