% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{min.symbolic_interval}
\alias{min.symbolic_interval}
\alias{max.symbolic_interval}
\alias{$.symbolic_interval}
\title{Maxima and Minima}
\usage{
\method{min}{symbolic_interval}(x, ...)

\method{max}{symbolic_interval}(x, ...)

\method{$}{symbolic_interval}(x, name = c("min", "max", "mean", "median"))
}
\arguments{
\item{x}{symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}

\item{name}{...}
}
\value{
a new symbolic interval with the minimum of the minima and the minimum of the maxima
}
\description{
Maxima and Minima
}
