% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.sym.data.table.R
\name{mean.sym.data.table}
\alias{mean.sym.data.table}
\title{Symbolic Mean}
\usage{
\method{mean}{sym.data.table}(x, method = c("centers", "interval", "modal"),
  trim = 0, na.rm = F, ...)
}
\arguments{
\item{x}{The symbolic variable.}

\item{method}{The method to be use.}

\item{trim}{As in R mean function.}

\item{na.rm}{As in R mean function.}

\item{...}{As in R mean function.}
}
\value{
Return a real number.
}
\description{
This function compute the symbolic mean
}
\examples{
data(example3)
sym.data<-example3
mean(sym.data[,1])
mean(sym.data[,2])
mean(sym.data[,2], method='interval')
mean(sym.data[,3], method='modal')

}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.

Rodriguez, O. (2000).
Classification et Modeles Lineaires en Analyse des Donnees Symboliques. Ph.D. Thesis,
Paris IX-Dauphine University.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Mean}
\keyword{Symbolic}
