% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{mean.symbolic_interval}
\alias{mean.symbolic_interval}
\title{Symbolic interval mean}
\usage{
\method{mean}{symbolic_interval}(x, ...)
}
\arguments{
\item{x}{a symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Symbolic interval mean
}
