% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{median.symbolic_interval}
\alias{median.symbolic_interval}
\title{Symbolic interval median}
\usage{
\method{median}{symbolic_interval}(x, ...)
}
\arguments{
\item{x}{a symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Symbolic interval median
}
