% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_regression.R
\name{method_summary}
\alias{method_summary}
\title{Summary method to CM and CRM regression model}
\usage{
method_summary(ref, pred)
}
\arguments{
\item{ref}{Real values}

\item{pred}{Predicted values}
}
\description{
Summary method to CM and CRM regression model
}
