% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{modal.distance}
\alias{modal.distance}
\title{Distance between modal multivalued variables}
\usage{
modal.distance(x, y, q = 2)
}
\arguments{
\item{x}{A symbilic data table 1x1}

\item{y}{A symbilic data table 1x1}
}
\description{
Distance between modal multivalued variables
}
\keyword{internal}
