% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{new.sym.histogram}
\alias{new.sym.histogram}
\title{Create an symbolic_histogram type object}
\usage{
new.sym.histogram(x = double(), breaks = NA_real_)
}
\description{
Create an symbolic_histogram type object
}
\keyword{internal}
