% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{new.sym.intreval}
\alias{new.sym.intreval}
\title{Create an symbolic_interval type object}
\usage{
new.sym.intreval(min = numeric(), max = numeric())
}
\description{
Create an symbolic_interval type object
}
\keyword{internal}
