% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{new_histogram}
\alias{new_histogram}
\title{Create an symbolic_histogram type object}
\usage{
new_histogram(x = double())
}
\description{
Create an symbolic_histogram type object
}
\keyword{internal}
