% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{new_sym_histogram}
\alias{new_sym_histogram}
\title{Create an symbolic_histogram type object}
\usage{
new_sym_histogram(x = double(), breaks = NA_real_)
}
\description{
Create an symbolic_histogram type object
}
\keyword{internal}
