% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{new_sym_intreval}
\alias{new_sym_intreval}
\title{Create an symbolic_interval type object}
\usage{
new_sym_intreval(min = numeric(), max = numeric())
}
\description{
Create an symbolic_interval type object
}
\keyword{internal}
