\name{norm.vect}
\alias{norm.vect}

\title{
Compute the norm of a vector
}
\description{
Compute the norm of a vector
}
\usage{
norm.vect(vector1)
}

\arguments{
  \item{vector1}{
An n dimensional vector.
}
}

\value{
The L2 norm of the vector.
}

\references{
Arce J. and Rodriguez O. (2015) "Principal Curves and Surfaces 
to Interval Valued Variables". The 5th Workshop on Symbolic 
Data Analysis, SDA2015, Orleans, France, November.

Hastie,T. (1984).
Principal Curves and Surface. Ph.D Thesis Stanford University.

Hastie,T. \& Weingessel,A. (2014).
princurve - Fits a Principal Curve in Arbitrary Dimension.R package version 1.1--12
http://cran.r-project.org/web/packages/princurve/index.html.

Hastie,T. \& Stuetzle, W. (1989). Principal Curves.
Journal of the American Statistical Association, Vol. 84-406, 502--516.

Hastie, T., Tibshirani, R. \& Friedman, J. (2008).
The Elements of Statistical Learning; Data Mining, Inference and Prediction. Springer, New York.
}
\author{
Jorge Arce.
}

\seealso{
\code{\link{sym.interval.pc}}
}

\keyword{ Principal }
\keyword{ Curve }
