% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.desv.fun.interval.j.r
\name{optim.desv.fun.interval}
\alias{optim.desv.fun.interval}
\title{optim.desv.fun.interval}
\usage{
optim.desv.fun.interval(sym.var.1)
}
\description{
optim.desv.fun.interval
}
\keyword{internal}
