% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.pca.distance.j.new.r
\name{optim.pca.distance.j.new}
\alias{optim.pca.distance.j.new}
\title{Optimized PCA Distance}
\usage{
optim.pca.distance.j.new(sym.data)
}
\arguments{
\item{sym.data}{An Interval Matrix}
}
\value{
Concept Projections onto the principal components,
        Classical PCA,
        Variance Best Matrix
}
\description{
Optimized PCA Distance
}
\keyword{internal}
