% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.pca.variance.j.new.r
\name{optim.pca.variance.j.new}
\alias{optim.pca.variance.j.new}
\title{Optimized PCA Variance}
\usage{
optim.pca.variance.j.new(sym.data, num.dimension)
}
\arguments{
\item{sym.data}{An Interval Matrix}

\item{num.dimension}{Number of dimensions}
}
\value{
Concept Projections onto the principal components,
        Classical PCA,
        Variance Best Matrix
}
\description{
Optimized PCA Variance
}
\keyword{internal}
