% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.to.sym.R
\name{pander.sym.data.table}
\alias{pander.sym.data.table}
\title{Pander method for symbolic data table}
\usage{
\method{pander}{sym.data.table}(x, caption = attr(x, "caption"), ...)
}
\arguments{
\item{x}{a symbolic data table}

\item{caption}{caption (string) to be shown under the table}

\item{...}{optional parameters passed to raw pandoc.table function}
}
\description{
Prints a symbolic data table in Pandoc's markdown
}
