% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.supplementary.vertex.fun.j.new.r
\name{pca.supplementary.vertex.fun.j.new}
\alias{pca.supplementary.vertex.fun.j.new}
\title{Calculate the distance}
\usage{
pca.supplementary.vertex.fun.j.new(
  x,
  N,
  M,
  sym.var.names,
  sym.data.vertex.matrix,
  tot.individuals
)
}
\arguments{
\item{x}{A Matrix}

\item{N}{Number of concepts}

\item{M}{Number of variables}

\item{sym.var.names}{Names of concepts}

\item{sym.data.vertex.matrix}{Vertex Matrix}

\item{tot.individuals}{Number of individuals}
}
\value{
Distance
}
\description{
Calculate the distance
}
