% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_pca.R
\name{pca.supplementary.vertex.fun.j}
\alias{pca.supplementary.vertex.fun.j}
\title{pca.supplementary.vertex.fun.j}
\usage{
pca.supplementary.vertex.fun.j(
  x,
  N,
  M,
  sym.var.names,
  sym.data.vertex.matrix,
  tot.individuals
)
}
\description{
pca.supplementary.vertex.fun.j
}
\keyword{internal}
