% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.supplementary.vertex.lambda.fun.j.r,
%   R/sym_interval_pca.R
\name{pca.supplementary.vertex.lambda.fun.j}
\alias{pca.supplementary.vertex.lambda.fun.j}
\title{Calculate the variance}
\usage{
pca.supplementary.vertex.lambda.fun.j(
  x,
  M,
  N,
  sym.var.names,
  sym.data.vertex.matrix,
  tot.individuals,
  num.dimen.aux
)

pca.supplementary.vertex.lambda.fun.j(
  x,
  M,
  N,
  sym.var.names,
  sym.data.vertex.matrix,
  tot.individuals,
  num.dimen.aux
)
}
\arguments{
\item{x}{A Matrix}

\item{M}{Number of variables}

\item{N}{Number of concepts}

\item{sym.var.names}{Names of concepts}

\item{sym.data.vertex.matrix}{Vertex Matrix}

\item{tot.individuals}{Number of individuals}

\item{num.dimen.aux}{Number of dimensions}
}
\value{
Cumulative variance
}
\description{
Calculate the variance

pca.supplementary.vertex.lambda.fun.j
}
\keyword{internal}
