% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.continuos.R
\name{plot.sym.continuos}
\alias{plot.sym.continuos}
\title{Title}
\usage{
\method{plot}{sym.continuos}(info, col = c("blue"), border = FALSE,
  show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
