% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.hist.R
\name{plot.sym.hist}
\alias{plot.sym.hist}
\title{Title}
\usage{
\method{plot}{sym.hist}(info, col = c("blue"), border = FALSE,
  ylab = TRUE, show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
