% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.interval.R
\name{plot.sym.interval}
\alias{plot.sym.interval}
\title{Title}
\usage{
\method{plot}{sym.interval}(info, col = c("blue"), border = FALSE,
  show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
\author{
Andres Navarro
}
