% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.modal.R
\name{plot.sym.modal}
\alias{plot.sym.modal}
\title{Title}
\usage{
\method{plot}{sym.modal}(info, col = c("blue"), border = FALSE,
  ylab = TRUE, show.type = TRUE, reduce = FALSE)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
