% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.row.R
\name{plot.sym.row}
\alias{plot.sym.row}
\title{Title}
\usage{
\method{plot}{sym.row}(data, col = NA, matrix.form = NA, border = FALSE,
  size = 1, title = TRUE, show.type = FALSE, reduce = FALSE)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
