% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.set.R
\name{plot.sym.set}
\alias{plot.sym.set}
\title{Title}
\usage{
\method{plot}{sym.set}(info, col = c("blue"), border = FALSE,
  show.type = TRUE, reduce = FALSE)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
