% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_tsne.R
\name{plot.sym_tsne}
\alias{plot.sym_tsne}
\title{Plot TSNE for symbolic data tables}
\usage{
\method{plot}{sym_tsne}(x, ...)
}
\arguments{
\item{x}{data frame}

\item{...}{other graphical parameters (see par and section ‘Details’ below).}
}
\description{
Plot TSNE for symbolic data tables
}
