% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_umap.R
\name{plot.sym_umap}
\alias{plot.sym_umap}
\title{Plot UMAP for symbolic data tables}
\usage{
\method{plot}{sym_umap}(x, ...)
}
\arguments{
\item{x}{sym_umap object}

\item{...}{params for plot}
}
\description{
Plot UMAP for symbolic data tables
}
