% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_pca.R
\name{plot.symbolic_pca}
\alias{plot.symbolic_pca}
\title{Plot symbolic PCA}
\usage{
\method{plot}{symbolic_pca}(x, choix = c("ind", "var"), axes = c(1, 2), labels = TRUE, ...)
}
\description{
Plot symbolic PCA
}
\keyword{internal}
