% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.to.sym.R
\name{print.sym.data.table}
\alias{print.sym.data.table}
\title{Printing Symbolic Data Table}
\usage{
\method{print}{sym.data.table}(x, ...)
}
\arguments{
\item{x}{Object of class sym.data.table}

\item{...}{optional arguments to print o format method}
}
\description{
Printing Symbolic Data Table
}
