% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.continuum.variable.R
\name{process.continuum.variable}
\alias{process.continuum.variable}
\title{process.continuum.variable}
\usage{
process.continuum.variable(variableName, conceptColumns)
}
\description{
process.continuum.variable
}
\keyword{internal}
