% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.histogram.variable.R
\name{process.histogram.variable}
\alias{process.histogram.variable}
\title{process.histogram.variable}
\usage{
process.histogram.variable(variableName, concept, dataTable, n)
}
\description{
process.histogram.variable
}
\keyword{internal}
