% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.interval.variable.R
\name{process.interval.variable}
\alias{process.interval.variable}
\title{process.interval.variable}
\usage{
process.interval.variable(variableName, conceptColumns)
}
\description{
process.interval.variable
}
\keyword{internal}
