% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.modal.variable.R
\name{process.modal.variable}
\alias{process.modal.variable}
\title{process.modal.variable}
\usage{
process.modal.variable(variableName, concept, sym.obj.names)
}
\description{
process.modal.variable
}
\keyword{internal}
