% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodas_to_rsda.R
\name{process.mult.nominal.modif.variable}
\alias{process.mult.nominal.modif.variable}
\title{process.mult.nominal.modif.variable}
\usage{
process.mult.nominal.modif.variable(
  labels,
  number.of.rows,
  parsed.xml,
  variable.index,
  variable.name
)
}
\description{
process.mult.nominal.modif.variable
}
\keyword{internal}
