% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.nominal.variable.R
\name{process.nominal.variable}
\alias{process.nominal.variable}
\title{process.nominal.variable}
\usage{
process.nominal.variable(labels, number.of.rows, parsed.xml, variable.index,
  variable.name)
}
\description{
process.nominal.variable
}
\keyword{internal}
