% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.set.variable.R
\name{process.set.variable}
\alias{process.set.variable}
\title{process.set.variable}
\usage{
process.set.variable(variableName, concept, sym.obj.names)
}
\description{
process.set.variable
}
\keyword{internal}
