% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.RSDA.r
\name{quantiles.RSDA}
\alias{quantiles.RSDA}
\title{quantiles.RSDA}
\usage{
quantiles.RSDA(histogram.matrix, num.quantiles)
}
\arguments{
\item{histogram.matrix}{A matrix of histograms}

\item{num.quantiles}{Number of quantiles}
}
\value{
Quantiles of a Histogram Matrix
}
\description{
quantiles.RSDA
}
\examples{
\dontrun{
data("hardwoodBrito")
Hardwood.histogram<-hardwoodBrito
Hardwood.cols<-colnames(Hardwood.histogram)
Hardwood.names<-row.names(Hardwood.histogram)
 M<-length(Hardwood.cols)
 N<-length(Hardwood.names)
 BIN.Matrix<-matrix(rep(3,N*M),nrow = N)
pca.hist<-sym.histogram.pca(Hardwood.histogram,BIN.Matrix)
Hardwood.quantiles.PCA<-quantiles.RSDA(pca.hist$sym.hist.matrix.PCA,3)
}
}
\author{
Jorge Arce Garro
}
