% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE.L}
\alias{RMSE.L}
\title{Lower boundary root-mean-square error}
\usage{
RMSE.L(ref, pred)
}
\arguments{
\item{ref}{Variable that was predicted.}

\item{pred}{The prediction given by the model.}
}
\value{
The lower boundary root-mean-square error.
}
\description{
Compute the lower boundary root-mean-square error.
}
\references{
LIMA-NETO, E.A., DE CARVALHO, F.A.T., (2008). Centre and range method
to fitting a linear regression model on symbolic interval data. Computational
Statistics and Data Analysis 52, 1500-1515.

LIMA-NETO, E.A., DE CARVALHO, F.A.T., (2010). Constrained linear regression models
for symbolic interval-valued variables. Computational Statistics and
Data Analysis 54, 333-347.
}
\seealso{
sym.glm
}
\author{
Oldemar Rodriguez Rojas.
}
\keyword{lower}
\keyword{root-mean-square}
