% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{select.dist}
\alias{select.dist}
\title{Select distance methods for a symbolic data table}
\usage{
select.dist(x)
}
\arguments{
\item{x}{A symbilic data table}
}
\description{
Select distance methods for a symbolic data table
}
\keyword{internal}
