% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{set.distance}
\alias{set.distance}
\title{Distance between set variables}
\usage{
set.distance(x, y, q = 2)
}
\arguments{
\item{x}{A symbilic data table 1x1}

\item{y}{A symbilic data table 1x1}

\item{method}{Method to use}

\item{normalize}{A logical value indicating whether normalize the output}
}
\description{
Distance between set variables
}
\keyword{internal}
