% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sodasXML}
\alias{sodasXML}
\title{sodasXML}
\format{An object of class \code{XMLInternalDocument} (inherits from \code{XMLAbstractDocument}) of length 1.}
\usage{
sodasXML
}
\description{
sodasXML
}
\keyword{internal}
