% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.data.r
\name{stand.data}
\alias{stand.data}
\title{Standardized Intervals}
\usage{
stand.data(sym.data, data.mean, data.stan, nn, mm)
}
\arguments{
\item{sym.data}{An Interval Matrix}

\item{data.mean}{A vector of means}

\item{data.stan}{A vector of standard deviation}

\item{nn}{Number of concepts}

\item{mm}{Number of variables}
}
\value{
Standardized intervals
}
\description{
Standardized Intervals
}
