\name{StudentsGrades}
\alias{StudentsGrades}
\docType{data}
\title{
Data Example
}
\description{
Symbolic data matrix with all the variables continuos type.
}
\usage{data(StudentsGrades)}
\format{
       $C Math $C Science $C Spanish $C History $C Sport
       
Lucia  $C  7.0 $C     6.5 $C     9.2 $C     8.6 $C   8.0

Pedro  $C  7.5 $C     9.4 $C     7.3 $C     7.0 $C   7.0

Ines   $C  7.6 $C     9.2 $C     8.0 $C     8.0 $C   7.5

Luis   $C  5.0 $C     6.5 $C     6.5 $C     7.0 $C   9.0

Andres $C  6.0 $C     6.0 $C     7.8 $C     8.9 $C   7.3

Ana    $C  7.8 $C     9.6 $C     7.7 $C     8.0 $C   6.5

Carlos $C  6.3 $C     6.4 $C     8.2 $C     9.0 $C   7.2

Jose   $C  7.9 $C     9.7 $C     7.5 $C     8.0 $C   6.0

Sonia  $C  6.0 $C     6.0 $C     6.5 $C     5.5 $C   8.7

Maria  $C  6.8 $C     7.2 $C     8.7 $C     9.0 $C   7.0	
}
\examples{
data(StudentsGrades)
StudentsGrades
}
\keyword{datasets}
