% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.data.table_subset.R
\name{[.sym.data.table}
\alias{[.sym.data.table}
\title{Extract or replace parts of a Symbolic Data Table}
\usage{
\method{[}{sym.data.table}(x, i, j)
}
\description{
Extract or replace parts of a Symbolic Data Table
}
\keyword{internal}
