% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.cfa.R
\name{sym.cfa}
\alias{sym.cfa}
\title{Symbolic correspondence analysis}
\usage{
sym.cfa(x = NULL, y  = NULL)
}
\arguments{
\item{x}{parametro x}

\item{y}{parametro y}
}
\value{
Return the interval principal components.
}
\description{
Correspondence Analysis for Symbolic MultiValued Variables.
}
\references{
Rodriguez, O. (2011).
Correspondence Analysis for Symbolic MultiValued Variables. Workshop in Symbolic Data Analysis
Namur, Belgium.
}
\author{
Oldemar Rodriguez Rojas.
}
\keyword{CFA}
