% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.continuos.gplot.R
\name{sym.continuos.gplot}
\alias{sym.continuos.gplot}
\title{sym.continuos.gplot}
\usage{
sym.continuos.gplot(info, color = c("black"), border = FALSE,
  show.type = TRUE, font.size = 1, background.color = "gray")
}
\description{
sym.continuos.gplot
}
\keyword{internal}
