% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.continuos.plot.R
\name{sym.continuos.plot}
\alias{sym.continuos.plot}
\title{Title}
\usage{
sym.continuos.plot(info, col = c("blue"), border = FALSE,
  show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
