% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.data.table.to.sym.var.R
\name{sym.data.table.to.sym.var}
\alias{sym.data.table.to.sym.var}
\title{sym.data.table.to.sym.var}
\usage{
sym.data.table.to.sym.var(sym.var)
}
\description{
sym.data.table.to.sym.var
}
\keyword{internal}
