% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_interval.R
\name{sym.dist.interval}
\alias{sym.dist.interval}
\title{Distance for Symbolic Interval Variables.}
\usage{
sym.dist.interval(
  sym.data,
  gamma = 0.5,
  method = "Minkowski",
  normalize = TRUE,
  SpanNormalize = FALSE,
  q = 1,
  euclidea = TRUE,
  pond = rep(1, length(variables))
)
}
\arguments{
\item{sym.data}{A symbolic object}

\item{gamma}{gamma value for the methods ichino and minkowski.}

\item{method}{Method to use (Gowda.Diday, Ichino, Minkowski, Hausdorff)}

\item{normalize}{A logical value indicating whether normalize the data in the ichino or hausdorff method.}

\item{SpanNormalize}{A logical value indicating whether}

\item{q}{q value for the hausdorff method.}

\item{euclidea}{A logical value indicating whether use the euclidean distance.}

\item{pond}{A numeric vector}

\item{variables}{Numeric vector with the number of the variables to use.}
}
\value{
An object of class 'dist'
}
\description{
This function computes and returns the distance matrix by using the specified
distance measure to compute distance between symbolic interval variables.
}
