% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_glm.R
\name{sym.glm}
\alias{sym.glm}
\title{Lasso, Ridge and and Elastic Net Linear regression model to interval variables}
\usage{
sym.glm(sym.data, response = 1, method = c('cm', 'crm'),
alpha = 1, nfolds = 10, grouped = TRUE)
}
\arguments{
\item{sym.data}{Should be a symbolic data table read with the function read.sym.table(...).}

\item{response}{The number of the column where is the response variable in the interval data table.}

\item{method}{'cm' to generalized Center Method and 'crm' to generalized Center and Range Method.}

\item{alpha}{alpha=1 is the lasso penalty, and alpha=0 the ridge penalty. 0<alpha<1 is the elastic net method.}

\item{nfolds}{Number of folds - default is 10. Although nfolds can be as large as the sample size
(leave-one-out CV), it is not recommended for large datasets. Smallest value allowable
is nfolds=3}

\item{grouped}{This is an experimental argument, with default TRUE, and can be ignored by most users.}
}
\value{
An object of class 'cv.glmnet' is returned, which is a list with the ingredients of the cross-validation fit.
}
\description{
Execute Lasso, Ridge and and Elastic Net Linear regression model to interval variables.
}
\references{
Rodriguez O. (2013). A generalization of Centre and Range method for fitting a linear
regression model to symbolic interval data using Ridge Regression, Lasso
and Elastic Net methods. The IFCS2013 conference of the International Federation of
Classification Societies, Tilburg University Holland.
}
\seealso{
sym.lm
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Lasso}
\keyword{Regression}
\keyword{Ridge}
\keyword{Symbolic}
