% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.hist.gplot.R
\name{sym.hist.gplot}
\alias{sym.hist.gplot}
\title{sym.hist.gplot}
\usage{
sym.hist.gplot(info, color = c("black"), border = FALSE, show.type = TRUE,
  fill = "white", background.color = "gray")
}
\description{
sym.hist.gplot
}
\keyword{internal}
