% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.hist.plot.R
\name{sym.hist.plot}
\alias{sym.hist.plot}
\title{sym.hist.plot}
\usage{
sym.hist.plot(info, col = c("blue"), border = FALSE, show.type = TRUE)
}
\description{
<<<<<<< HEAD
sym.hist.plot
=======
Function for plotting one "histogram" type value from the symbolic data table
}
\examples{
\dontrun{
data(ex1_db2so)
data.sym <- classic.to.sym(ex1_db2so, concept=c("state", "sex"),
                           col.names = c(county, group, age,age,age,age),
                           variables.types=c(county = type.interval(),
                           group = type.continuous(),
                           age = type.continuous(),
                           age = type.continuous(),
                           age = type.modal(),
                           age = type.histogram()))
sym.hist.plot(data.sym[1,6])
}
}
\author{
Andres Navarro
>>>>>>> master
}
\keyword{internal}
