% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{sym.histogram}
\alias{sym.histogram}
\title{Create an symbolic_histogram type object}
\usage{
sym.histogram(x = double(), breaks = NA_real_)
}
\arguments{
\item{x}{character vector}

\item{breaks}{a vector giving the breakpoints between histogram cells}
}
\value{
a symbolic histogram
}
\description{
Create an symbolic_histogram type object
}
\examples{
sym.histogram(iris$Sepal.Length)
}
