% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.histogram.to.latex.r
\name{sym.histogram.to.latex}
\alias{sym.histogram.to.latex}
\title{optim.desv.fun.interval}
\usage{
sym.histogram.to.latex(datasym.V3.histogram)
}
\description{
optim.desv.fun.interval
}
\keyword{internal}
