% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_interval.R
\name{sym.Interval.distance}
\alias{sym.Interval.distance}
\title{Compute the distance between two rows}
\usage{
sym.Interval.distance(
  sym.data,
  variable,
  w1,
  w2,
  gamma = 0.5,
  method = "Minkowski",
  normalize = TRUE
)
}
\description{
Compute the distance between two rows
}
\keyword{internal}
