% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.interval.gplot.R
\name{sym.interval.gplot}
\alias{sym.interval.gplot}
\title{sym.interval.gplot}
\usage{
sym.interval.gplot(info, color = c("black"), border = FALSE,
  show.type = TRUE, fill = "white", font.size = 1,
  background.color = "gray")
}
\description{
sym.interval.gplot
}
\keyword{internal}
