% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.interval.pc.2.j.r
\name{sym.interval.pc.2.j}
\alias{sym.interval.pc.2.j}
\title{sym.interval.pc.2.j}
\usage{
sym.interval.pc.2.j(sym.data, method = c("vertex", "centers"), maxit, plot,
  scale, center)
}
\description{
sym.interval.pc.2.j
}
\keyword{internal}
