% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.interval.pc.limits.2.j.r
\name{sym.interval.pc.limits.2.j}
\alias{sym.interval.pc.limits.2.j}
\title{sym.interval.pc.limits.2.j}
\usage{
sym.interval.pc.limits.2.j(sym.data, prin.curve, num.vertex, lambda)
}
\description{
sym.interval.pc.limits.2.j
}
\keyword{internal}
