% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_pca_limits_new_j.R
\name{sym.interval.pca.limits.new.j}
\alias{sym.interval.pca.limits.new.j}
\title{sym.interval.pca.limits.new.j}
\usage{
sym.interval.pca.limits.new.j(sym.data, prin.comp, num.vertex)
}
\description{
sym.interval.pca.limits.new.j
}
\keyword{internal}
