% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.interval.plot.R
\name{sym.interval.plot}
\alias{sym.interval.plot}
\title{Title}
\usage{
sym.interval.plot(info, col = c("blue"), border = FALSE, show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
\author{
Andres Navarro
}
