% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_pca.R
\name{sym.interval.vertex.pca.j}
\alias{sym.interval.vertex.pca.j}
\title{sym.interval.vertex.pca.j}
\usage{
sym.interval.vertex.pca.j(data.sym)
}
\description{
sym.interval.vertex.pca.j
}
\keyword{internal}
