% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.modal.gplot.R
\name{sym.modal.gplot}
\alias{sym.modal.gplot}
\title{sym.modal.gplot}
\usage{
sym.modal.gplot(info, color = c("black"), border = FALSE,
  show.type = TRUE, reduce = FALSE, fill = "white",
  background.color = "gray")
}
\description{
sym.modal.gplot
}
\keyword{internal}
