% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{sym.modal}
\alias{sym.modal}
\title{Create an symbolic_modal type object}
\usage{
sym.modal(x = character())
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic modal
}
\description{
Create an symbolic_modal type object
}
\examples{
sym.modal(factor(c("a", "b", "b", "l")))
}
