% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.set.R
\name{sym.Multieval.distance}
\alias{sym.Multieval.distance}
\title{Compute the distance between two rows}
\usage{
sym.Multieval.distance(sym.data, variable, w1, w2, gamma = 0.5,
  method = "Minkowski", normalize = TRUE)
}
\description{
Compute the distance between two rows
}
\keyword{internal}
