% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_regression.R
\name{sym.predict.symbolic_gbm_cm}
\alias{sym.predict.symbolic_gbm_cm}
\title{Predict model_gbm_cm model}
\usage{
\method{sym.predict}{symbolic_gbm_cm}(model, new.sym.data, n.trees = 500, ...)
}
\arguments{
\item{model}{model}

\item{new.sym.data}{new data}

\item{n.trees}{Integer specifying the total number of trees to fit. This is equivalent to the number of iterations and the number of basis functions in the additive expansion. Default is 100.}

\item{...}{optional parameters}
}
\description{
Predict model_gbm_cm model
}
\references{
Lima-Neto, E.A., De Carvalho, F.A.T., (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis52, 1500-1515


Lima-Neto, E.A., De Carvalho, F.A.T., (2010). Constrained linear regression models for symbolic interval-valued variables.  Computational Statistics and Data Analysis 54, 333-347


Lima Neto, E.d.A., de Carvalho, F.d.A.T. Nonlinear regression applied to interval-valued data. Pattern Anal Applic 20, 809–824 (2017). https://doi.org/10.1007/s10044-016-0538-y


Rodriguez, O. (2018). Shrinkage linear regression for symbolic interval-valued variables.Journal MODULAD 2018, vol. Modulad 45, pp.19-38
}
