% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_radar_plot.R
\name{sym.radar.data}
\alias{sym.radar.data}
\title{Internal sym.radar.data}
\usage{
sym.radar.data(dat, nom.vars, use.pct = F)
}
\arguments{
\item{dat}{The symbolic data.}

\item{nom.vars}{Names of the variables.}

\item{use.pct}{a logical value indicating if use percentage or real distance for plot.}
}
\value{
A molten data.
}
\description{
Internal sym.radar.data
}
\keyword{Molten}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{plot}
\keyword{radar}
