% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.radar.plot.R
\name{sym.radar.plot.}
\alias{sym.radar.plot.}
\title{Internal sym.radar.plot the distence between two rows}
\usage{
sym.radar.plot.(dat, indivs, vars, dat.min, dat.max, rad.main, seg = 3,
  use.legend = T)
}
\description{
Internal sym.radar.plot the distence between two rows
}
\keyword{internal}
