% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.radar.plot.R
\name{sym.radar.plot}
\alias{sym.radar.plot}
\title{Radar Plot For Symbolic Interval Variables}
\usage{
sym.radar.plot(x, dat.min, dat.max, rad.main, seg = 3, use.legend = T,
  main = "")
}
\arguments{
\item{x}{A symbolic object.}

\item{dat.min}{<documentar>}

\item{dat.max}{<documentar>}

\item{rad.main}{<documentar>}

\item{seg}{<documentar>}

\item{use.legend}{a logical value indicating whether use a legend.}

\item{main}{the main title (on top).}
}
\value{
A radar plot.
}
\description{
Radar Plot For Symbolic Interval Variables
}
