% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_regression.R
\name{sym.rf}
\alias{sym.rf}
\title{Symbolic Regression with Random Forest}
\usage{
sym.rf(formula, sym.data, method = c("cm", "crm"), ntree = 500)
}
\arguments{
\item{formula}{a formula, with a response but no interaction terms. If this a a data frame, that is taken as the model frame (see model.frame).}

\item{sym.data}{symbolic data table}

\item{method}{cm crm}

\item{ntree}{Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}
}
\description{
Symbolic Regression with Random Forest
}
\references{
Lima-Neto, E.A., De Carvalho, F.A.T., (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis52, 1500-1515


Lima-Neto, E.A., De Carvalho, F.A.T., (2010). Constrained linear regression models for symbolic interval-valued variables.  Computational Statistics and Data Analysis 54, 333-347


Lima Neto, E.d.A., de Carvalho, F.d.A.T. Nonlinear regression applied to interval-valued data. Pattern Anal Applic 20, 809–824 (2017). https://doi.org/10.1007/s10044-016-0538-y


Rodriguez, O. (2018). Shrinkage linear regression for symbolic interval-valued variables.Journal MODULAD 2018, vol. Modulad 45, pp.19-38
}
