% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.scale.matrix.r
\name{sym.scale.matrix.j}
\alias{sym.scale.matrix.j}
\title{sym.scale.matrix.j}
\usage{
sym.scale.matrix.j(data, data.mean, data.sd, nn, mm)
}
\arguments{
\item{data}{A symbolic data}

\item{data.mean}{A vector of means}

\item{data.sd}{A vector of deviation standard}

\item{nn}{Number of concepts}

\item{mm}{Number of variables}
}
\value{
Standardized Data
}
\description{
sym.scale.matrix.j
}
\keyword{internal}
