% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.set.gplot.R
\name{sym.set.gplot}
\alias{sym.set.gplot}
\title{sym.set.gplot}
\usage{
sym.set.gplot(info, color = c("black"), border = FALSE, show.type = TRUE,
  reduce = FALSE, fill = "white", background.color = "gray")
}
\description{
sym.set.gplot
}
\keyword{internal}
