% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{sym.set}
\alias{sym.set}
\title{Create an symbolic_set type object}
\usage{
sym.set(x = NA)
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic set
}
\description{
Create an symbolic_set type object
}
\examples{
sym.set(factor(c("a", "b", "b", "l")))
}
