% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_histogram.R
\name{sym_histogram}
\alias{sym_histogram}
\title{Create an symbolic_histogram type object}
\usage{
sym_histogram(x = double(), breaks = NA_real_)
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic histogram
}
\description{
Create an symbolic_histogram type object
}
\examples{
sym_histogram(iris$Sepal.Length)
}
