% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{sym_interval}
\alias{sym_interval}
\title{Create an symbolic_interval type object}
\usage{
sym_interval(x = double(), .min = min, .max = max)
}
\arguments{
\item{x}{numeric vector}

\item{.min}{function that will be used to calculate the minimum interval}

\item{.max}{function that will be used to calculate the maximum interval}
}
\value{
a symbolic interval
}
\description{
Create an symbolic_interval type object
}
\examples{
sym_interval(c(1,2,4,5))
sym_interval(1:10)
}
