% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_modal.R
\name{sym_modal}
\alias{sym_modal}
\title{Create an symbolic_modal type object}
\usage{
sym_modal(x = character())
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic modal
}
\description{
Create an symbolic_modal type object
}
\examples{
sym_modal(c("a","b","b","l"))
}
