% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_set.R
\name{sym_set}
\alias{sym_set}
\title{Create an symbolic_set type object}
\usage{
sym_set(x = character())
}
\arguments{
\item{x}{character vector}
}
\value{
a symbolic set
}
\description{
Create an symbolic_set type object
}
\examples{
sym_set(c("a","b","b","l"))
}
