% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{sd}
\alias{sd}
\alias{sd.default}
\alias{sd.symbolic_interval}
\title{Symbolic interval SD}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)

\method{sd}{symbolic_interval}(x, ...)
}
\arguments{
\item{x}{a symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Symbolic interval SD
}
