% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_interval.R
\name{var}
\alias{var}
\title{Symbolic interval variance}
\usage{
var(x, ...)
}
\arguments{
\item{x}{a symbolic interval vector}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Symbolic interval variance
}
