% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{mean.symbolic_interval}
\alias{mean.symbolic_interval}
\alias{mean.symbolic_tbl}
\title{Symbolic mean for intervals}
\usage{
\method{mean}{symbolic_interval}(x, method = c("centers", "interval"), trim = 0, na.rm = F, ...)

\method{mean}{symbolic_tbl}(x, ...)
}
\arguments{
\item{x}{A symbolic interval.}

\item{method}{The method to be use.}

\item{trim}{As in R mean function.}

\item{na.rm}{As in R mean function.}

\item{...}{As in R mean function.}
}
\description{
This function compute the symbolic mean for intervals
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.

Rodriguez, O. (2000).
Classification et Modeles Lineaires en Analyse des Donnees Symboliques. Ph.D. Thesis,
Paris IX-Dauphine University.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Mean}
\keyword{Symbolic}
