% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{median.symbolic_interval}
\alias{median.symbolic_interval}
\alias{median.symbolic_tbl}
\title{Symbolic Median}
\usage{
\method{median}{symbolic_interval}(x, na.rm = FALSE, method = c("centers", "interval"), ...)

\method{median}{symbolic_tbl}(x, ...)
}
\arguments{
\item{x}{A symbolic interval.}

\item{na.rm}{As in R median function.}

\item{method}{The method to be use.}

\item{...}{As in R median function.}
}
\description{
This function compute the median for symbolic intervals.
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.

Rodriguez, O. (2000).
Classification et Modeles Lineaires en Analyse des Donnees Symboliques. Ph.D. Thesis,
Paris IX-Dauphine University.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Median}
\keyword{Symbolic}
