% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table7}
\alias{Table7}
\alias{table7}
\title{Table7}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
table7
}
\description{
example for the dist.interval function.
}
\keyword{datasets}
