\name{USCrime}
\alias{USCrime}
\docType{data}
\title{
Us crime classic data table
}
\description{
Us crime classic data table that can be used to generate symbolic data tables.
}
\usage{data(USCrime)}

\source{
http://archive.ics.uci.edu/ml/
}
\references{
HASTIE, T., TIBSHIRANI, R. and FRIEDMAN, J. (2008). The Elements of Statistical Learning: 
Data Mining, Inference and Prediction. New York: Springer.
}
\examples{
data(USCrime)
us.crime<-USCrime
dim(us.crime)
head(us.crime)
summary(us.crime)
names(us.crime)
nrow(us.crime)
concept<-Sconcept(us.crime$state)
us.crime.int<-Sobject(Sint(concept$fold),Sint(concept$population),
Sint(concept$householdsize),Sint(concept$racepctblack),
Sint(concept$racePctWhite),Sint(concept$racePctAsian),
Sint(concept$racePctHisp),Sint(concept$agePct12t21),
Sint(concept$agePct12t29),Sint(concept$agePct16t24),
Sint(concept$agePct65up),Sint(concept$numbUrban),
Sint(concept$pctUrban),Sint(concept$medIncome),
Sint(concept$pctWWage),Sint(concept$pctWFarmSelf),
Sint(concept$pctWInvInc),Sint(concept$pctWSocSec),
Sint(concept$pctWPubAsst),Sint(concept$pctWRetire),
Sint(concept$medFamInc),Sint(concept$perCapInc),
Sint(concept$whitePerCap),Sint(concept$blackPerCap),
Sint(concept$indianPerCap),Sint(concept$AsianPerCap),
Sint(concept$OtherPerCap),Sint(concept$HispPerCap),
Sint(concept$NumUnderPov),
Sint(concept$PctPopUnderPov),Sint(concept$PctLess9thGrade),
Sint(concept$PctNotHSGrad),Sint(concept$PctBSorMore),
Sint(concept$PctUnemployed),Sint(concept$PctEmploy),
Sint(concept$PctEmplManu),Sint(concept$PctEmplProfServ),
Sint(concept$PctOccupManu),Sint(concept$PctOccupMgmtProf),
Sint(concept$MalePctDivorce),Sint(concept$MalePctNevMarr),
Sint(concept$FemalePctDiv),Sint(concept$TotalPctDiv),
Sint(concept$PersPerFam),Sint(concept$PctFam2Par),
Sint(concept$PctKids2Par),Sint(concept$PctYoungKids2Par),
Sint(concept$PctTeen2Par),Sint(concept$PctWorkMomYoungKids),
Sint(concept$PctWorkMom),Sint(concept$NumIlleg),
Sint(concept$PctIlleg),Sint(concept$NumImmig),
Sint(concept$PctImmigRecent),Sint(concept$PctImmigRec5),
Sint(concept$PctImmigRec8),Sint(concept$PctImmigRec10),
Sint(concept$PctRecentImmig),Sint(concept$PctRecImmig5),
Sint(concept$PctRecImmig8),Sint(concept$PctRecImmig10),
Sint(concept$PctSpeakEnglOnly),Sint(concept$ PctNotSpeakEnglWell),
Sint(concept$PctLargHouseFam),Sint(concept$PctLargHouseOccup),
Sint(concept$PersPerOccupHous),Sint(concept$PersPerOwnOccHous),
Sint(concept$PersPerRentOccHous),Sint(concept$PctPersOwnOccup),
Sint(concept$PctPersDenseHous),Sint(concept$PctHousLess3BR),
Sint(concept$MedNumBR),Sint(concept$HousVacant),
Sint(concept$PctHousOccup),Sint(concept$PctHousOwnOcc),
Sint(concept$PctVacantBoarded),Sint(concept$PctVacMore6Mos),
Sint(concept$MedYrHousBuilt),Sint(concept$PctHousNoPhone),
Sint(concept$PctWOFullPlumb),Sint(concept$OwnOccLowQuart),
Sint(concept$OwnOccMedVal),Sint(concept$OwnOccHiQuart),
Sint(concept$RentLowQ),Sint(concept$RentMedian),
Sint(concept$RentHighQ),Sint(concept$MedRent),
Sint(concept$MedRentPctHousInc),Sint(concept$MedOwnCostPctInc),
Sint(concept$MedOwnCostPctIncNoMtg),Sint(concept$NumInShelters),
Sint(concept$NumStreet),Sint(concept$PctForeignBorn),
Sint(concept$PctBornSameState),Sint(concept$PctSameHouse85),
Sint(concept$PctSameCity85),Sint(concept$PctSameState85),
Sint(concept$LandArea),Sint(concept$PopDens),
Sint(concept$PctUsePubTrans),Sint(concept$LemasPctOfficDrugUn),
Sint(concept$ViolentCrimesPerPop))
}
\keyword{datasets}
