% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{vec_ptype_abbr.symbolic_modal}
\alias{vec_ptype_abbr.symbolic_modal}
\title{abbr for symbolic modal}
\usage{
vec_ptype_abbr.symbolic_modal(x)
}
\description{
abbr for symbolic modal
}
\keyword{internal}
