% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{vec_ptype_full.symbolic_histogram}
\alias{vec_ptype_full.symbolic_histogram}
\title{full name for symbolic modal}
\usage{
vec_ptype_full.symbolic_histogram(x)
}
\description{
full name for symbolic modal
}
\keyword{internal}
