% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{vec_ptype_full.symbolic_interval}
\alias{vec_ptype_full.symbolic_interval}
\title{full name for symbolic interval}
\usage{
vec_ptype_full.symbolic_interval(x)
}
\description{
full name for symbolic interval
}
\keyword{internal}
