% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_objects.R
\name{vec_ptype_full.symbolic_modal}
\alias{vec_ptype_full.symbolic_modal}
\title{full name for symbolic set}
\usage{
vec_ptype_full.symbolic_modal(x)
}
\description{
full name for symbolic set
}
\keyword{internal}
