% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.center.Hist.RSDA.r
\name{weighted.center.Hist.RSDA}
\alias{weighted.center.Hist.RSDA}
\title{weighted.center.Hist.RSDA}
\usage{
weighted.center.Hist.RSDA(sym.histogram)
}
\arguments{
\item{sym.histogram}{A Histogram matrix}
}
\value{
Matrix of Weighted Centers
}
\description{
weighted.center.Hist.RSDA
}
\examples{
\dontrun{
data(hardwoodBrito)
weighted.center.Hist.RSDA(hardwoodBrito)
}
}
\author{
Jorge Arce Garro
}
