% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{SODAS XML data file.}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 7 columns.
}
\source{
http://www.info.fundp.ac.be/asso/sodaslink.htm
}
\usage{
data(abalone)
}
\description{
Example of SODAS XML data file converted in a CSV file in RSDA format.
}
\examples{
data(abalone)
res <- sym.pca(abalone, 'centers')
plot(res, choix = "ind")
plot(res, choix = "var")
}
\references{
Bock H-H. and Diday E. (eds.) (2000).Analysis of Symbolic Data. Exploratory methods
for extracting statistical information fromcomplex data. Springer, Germany.
}
\keyword{datasets}
