% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cardiological}
\alias{Cardiological}
\title{Cardiological data example}
\format{An object of class \code{list} of length 9.}
\usage{
data(Cardiological)
}
\description{
Cardiological interval data example.
}
\examples{
data(Cardiological)
car.data <- Cardiological
res.cm <- sym.lm(Pulse~Syst+Diast,sym.data=car.data,method='cm')
pred.cm <- predictsym.lm(res.cm,car.data,method='cm')
RMSE.L(sym.var(car.data,1),pred.cm$Fitted)
RMSE.U(sym.var(car.data,1),pred.cm$Fitted)
R2.L(sym.var(car.data,1),pred.cm$Fitted)
R2.U(sym.var(car.data,1),pred.cm$Fitted)
deter.coefficient(sym.var(car.data,1),pred.cm$Fitted)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
