% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cardiological}
\alias{Cardiological}
\title{Cardiological data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.
}
\usage{
data(Cardiological)
}
\description{
Cardiological interval data example.
}
\examples{
data(Cardiological)
res.cm <- sym.lm(formula = Pulse~Syst+Diast, sym.data = Cardiological, method = 'cm')
pred.cm <- sym.predict(res.cm, Cardiological)
RMSE.L(Cardiological$Pulse, pred.cm$Fitted)
RMSE.U(Cardiological$Pulse,pred.cm$Fitted)
R2.L(Cardiological$Pulse,pred.cm$Fitted)
R2.U(Cardiological$Pulse,pred.cm$Fitted)
deter.coefficient(Cardiological$Pulse,pred.cm$Fitted)
}
\references{
Billard L. and  Diday E. (2006).Symbolic data analysis:
Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
