% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.to.RSDA.inteval.table.j.r
\name{data.frame.to.RSDA.inteval.table.j}
\alias{data.frame.to.RSDA.inteval.table.j}
\title{data.frame.to.RSDA.inteval.table.j}
\usage{
data.frame.to.RSDA.inteval.table.j(data.df)
}
\description{
data.frame.to.RSDA.inteval.table.j
}
\keyword{internal}
