% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.sym.table.R
\name{display.sym.table}
\alias{display.sym.table}
\title{display.sym.table}
\usage{
display.sym.table(sym.data)
}
\arguments{
\item{sym.data}{Shoud be a Symbolic Data table that have been read with read.sym.table(...)}
}
\value{
Not value.
}
\description{
This function display a symbolic data table tha have been read by read.sym.table(...)
}
\details{
The output will be the symbolic data table in the screen.
}
\examples{
data(example3)
display.sym.table(example3)
}
\references{
Billard, L and  Diday, E. (2007). Symbolic Data Analysis: Conceptual Statistics and Data
Mining (Wiley Series in Computational Statistics).
Billard, L and  Diday, E. (2003). From the Statistics of Data to the Statistics of
Knowledge: Symbolic Data Analysis. Journal of the American of the Statistical Association, USA.
}
\seealso{
read.sym.table
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Display}
\keyword{Symbolic}
\keyword{Table}
