% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_cfa2}
\alias{ex_cfa2}
\title{Correspondence Analysis Example}
\format{An object of class \code{list} of length 9.}
\usage{
data(ex_cfa2)
}
\description{
Correspondence Analysis for Symbolic MultiValued Variables example.
}
\references{
Rodriguez, O. (2011).
Correspondence Analysis for Symbolic MultiValued Variables. Workshop in Symbolic
Data Analysis Namur, Belgium
}
\keyword{datasets}
