% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_cfa2}
\alias{ex_cfa2}
\title{Correspondence Analysis Example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
data(ex_cfa2)
}
\description{
Correspondence Analysis for Symbolic MultiValued Variables example.
}
\references{
Rodriguez, O. (2011).
Correspondence Analysis for Symbolic MultiValued Variables. Workshop in Symbolic
Data Analysis Namur, Belgium
}
\keyword{datasets}
