% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_mcfa1}
\alias{ex_mcfa1}
\title{Multiple Correspondence Analysis Example}
\format{
An object of class \code{data.frame} with 130 rows and 5 columns.

An object of class \code{data.frame} with 130 rows and 5 columns.
}
\usage{
data(ex_mcfa1)

ex_mcfa1
}
\description{
example for the sym.mcfa function.

example for the sym.mcfa function.
}
\examples{
data("ex_mcfa1")
sym.table <- classic.to.sym(ex_mcfa1,
                            concept = suspect,
                            hair = sym.set(hair),
                            eyes = sym.set(eyes),
                            region = sym.set(region))

res <- sym.mcfa(sym.table, c(1,2))
mcfa.scatterplot(res[,1], res[,2], sym.data = sym.table, pos.var = c(1,2))

data("ex_mcfa1")
sym.table <- classic.to.sym(
  x = ex_mcfa1,
  concept = "suspect",
  variables = c(hair, eyes, region),
  hair = sym.set(hair),
  eyes = sym.set(eyes),
  region = sym.set(region)
)
sym.table
}
\keyword{datasets}
