% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_mcfa2}
\alias{ex_mcfa2}
\title{Multiple Correspondence Analysis Example}
\format{An object of class \code{data.frame} with 130 rows and 7 columns.}
\usage{
ex_mcfa2
}
\description{
example for the sym.mcfa function.
}
\examples{
data("ex_mcfa2")

ex <- classic.to.sym(ex_mcfa2, concept = "employee_id", variables = everything(),
                    variables.types = c(salary = type.set(),
                                        region = type.set(),
                                        evaluation = type.set(),
                                        years_worked = type.set()))

res <- sym.mcfa(ex, c(1,2,3,4))
mcfa.scatterplot(res[,1], res[,2], sym.data = ex, pos.var = c(1,2,3,4))

}
\keyword{datasets}
