\name{example6}
\alias{example6}
\docType{data}
\title{
Data Example 7
}
\description{
This a symbolic data matrix wint continuos, interval, histograma a set data types.
}
\usage{data(example6)}
\format{
$C   F1 $H F2  M1  M2  M3  M4  M5 $I F3 F3 $H F4  M1  M2  M3 $C   F5 $S F6 E1 E2 E3 E4

Case1 $C  2.8 $H  5 0.1 0.1 0.1 0.1 0.6 $I  1  2 $H  3 0.1 0.7 0.2 $C  6.0 $S  4  e  g  k  i

Case2 $C  1.4 $H  5 0.1 0.1 0.1 0.1 0.6 $I  3  9 $H  3 0.6 0.3 0.1 $C  8.0 $S  4  a  b  c  d

Case3 $C  3.2 $H  5 0.1 0.1 0.1 0.1 0.6 $I -1  4 $H  3 0.2 0.2 0.6 $C -7.0 $S  4  2  1  b  c

Case4 $C -2.1 $H  5 0.1 0.1 0.1 0.1 0.6 $I  0  2 $H  3 0.9 0.0 0.1 $C  0.0 $S  4  3  4  c  a

Case5 $C -3.0 $H  5 0.1 0.1 0.1 0.1 0.6 $I -4 -2 $H  3 0.6 0.0 0.4 $C -9.5 $S  4  e  i  g  k
 }

\examples{
data(example6)
display.sym.table(example6)
}
\keyword{datasets}
