% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.columns.multivalued.j.r
\name{generate.columns.multivalued}
\alias{generate.columns.multivalued}
\title{generate.columns.multivalued}
\usage{
generate.columns.multivalued(data)
}
\description{
generate.columns.multivalued
}
\keyword{internal}
