% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.sym.table.r
\name{generate.sym.table}
\alias{generate.sym.table}
\title{generate.sym.table}
\usage{
generate.sym.table(sym.data)
}
\description{
generate.sym.table
}
\author{
Jorge Arce Garro
}
\keyword{internal}
