% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lynne1}
\alias{lynne1}
\title{Symbolic interval data example.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
data(lynne1)
}
\description{
Symbolic data matrix with all the variables of interval type.
}
\examples{
data(lynne1)
lynne1
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
