% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lynne2}
\alias{lynne2}
\title{Symbolic interval data example.}
\format{An object of class \code{sym.data.table} of length 9.}
\usage{
data(lynne2)
}
\description{
Symbolic data matrix with all the variables of interval type.
}
\examples{
data(lynne2)
display.sym.table(lynne2)
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
