% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.data.table.R
\name{plot.sym.data.table}
\alias{plot.sym.data.table}
\title{Title}
\usage{
\method{plot}{sym.data.table}(x, col = NA, border = FALSE, size = 1,
  title = TRUE, show.type = FALSE, reduce = FALSE, ...)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
\keyword{Plot}
\keyword{Symbolic}
\keyword{data}
\keyword{table}
