% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodas_to_rsda.R
\name{process.continue.variable}
\alias{process.continue.variable}
\title{process.continue.variable}
\usage{
process.continue.variable(
  number.of.rows,
  parsed.xml,
  variable.index,
  variable.name
)
}
\description{
process.continue.variable
}
\keyword{internal}
