% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StudentsGrades}
\alias{StudentsGrades}
\title{Data Example.}
\format{$C Math  $C Science  $C Spanish  $C History  $C Sport\cr
Lucia   $C  7.0   $C     6.5  $C     9.2  $C     8.6  $C   8.0\cr
Pedro   $C  7.5  $C     9.4  $C     7.3  $C     7.0  $C   7.0\cr
Ines    $C  7.6  $C     9.2  $C     8.0  $C     8.0  $C   7.5\cr
Luis    $C  5.0  $C     6.5  $C     6.5  $C     7.0  $C   9.0\cr
Andres  $C  6.0  $C     6.0  $C     7.8  $C     8.9  $C   7.3\cr
Ana     $C  7.8  $C     9.6  $C     7.7  $C     8.0  $C   6.5\cr
Carlos  $C  6.3  $C     6.4  $C     8.2  $C     9.0  $C   7.2\cr
Jose    $C  7.9  $C     9.7  $C     7.5  $C     8.0  $C   6.0\cr
Sonia   $C  6.0  $C     6.0  $C     6.5  $C     5.5  $C   8.7\cr
Maria   $C  6.8  $C     7.2  $C     8.7  $C     9.0  $C   7.0\cr}
\usage{
data(StudentsGrades)
}
\description{
Symbolic data matrix with all the variables continuos type.
}
\examples{
data(StudentsGrades)
StudentsGrades
}
\keyword{datasets}
