% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_circle_plot.R
\name{sym.circle.plot}
\alias{sym.circle.plot}
\title{Symbolic Circle of Correlations}
\usage{
sym.circle.plot(prin.corre)
}
\arguments{
\item{prin.corre}{A symbolic interval data matrix with correlations between the variables and the
principals componets, both of interval type.}
}
\value{
Plot the symbolic circle
}
\description{
Plot the symbolic circle of correlations.
}
\examples{
data(oils)
res <- sym.pca(oils, "centers")
sym.circle.plot(res$Sym.Prin.Correlations)
}
\references{
Rodriguez O. (2012). The Duality Problem in Interval Principal Components Analysis.
The 3rd Workshop in Symbolic Data Analysis, Madrid.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Circle}
\keyword{Symbolic}
