% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.symbolic_df.R
\name{sym.continuos.plot}
\alias{sym.continuos.plot}
\title{sym.continuos.plot}
\usage{
sym.continuos.plot(info, col = c("blue"), border = FALSE, show.type = TRUE)
}
\description{
sym.continuos.plot
}
\keyword{internal}
