% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.symbolic_df.R
\name{sym.hist.plot}
\alias{sym.hist.plot}
\title{sym.hist.plot}
\usage{
sym.hist.plot(
  info,
  col = c("blue"),
  border = FALSE,
  show.type = TRUE,
  angle = 60
)
}
\description{
sym.hist.plot
}
\keyword{internal}
