% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.hist.plot.R
\name{sym.hist.plot}
\alias{sym.hist.plot}
\title{Title}
\usage{
sym.hist.plot(info, col = c("blue"), border = FALSE, ylab = TRUE,
  show.type = TRUE)
}
\arguments{
\item{show.type}{}
}
\description{
Title
}
