% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.histogram.pca.r
\name{sym.histogram.pca}
\alias{sym.histogram.pca}
\title{sym.histogram.pca}
\usage{
sym.histogram.pca(sym.hist.matrix, BIN.Matrix, method = NULL)
}
\arguments{
\item{sym.hist.matrix}{A Histogram matrix}

\item{BIN.Matrix}{A matrix with the number of bins for each individual and variable}

\item{method}{Weigthed Method}
}
\value{
Histogram PCA
}
\description{
sym.histogram.pca
}
\examples{
\dontrun{
data("hardwoodBrito")
Hardwood.histogram<-hardwoodBrito
weighted.center<-weighted.center.Hist.RSDA(Hardwood.histogram)
 M<-length(Hardwood.cols)
 N<-length(Hardwood.names)
 BIN.Matrix<-matrix(rep(3,N*M),nrow = N)
pca.hist<-sym.histogram.pca(Hardwood.histogram,BIN.Matrix)
pca.hist
}

}
\author{
Jorge Arce Garro
}
