% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_pc_limits.R
\name{sym.interval.pc.limits}
\alias{sym.interval.pc.limits}
\title{Symbolic interval principal curves limits}
\usage{
sym.interval.pc.limits(sym.data, prin.curve, num.vertex, lambda, var.ord)
}
\arguments{
\item{sym.data}{Symbolic interval data table.}

\item{prin.curve}{Principal curves.}

\item{num.vertex}{Number of vertices of the hipercube.}

\item{lambda}{Lambda.}

\item{var.ord}{Order of the variables.}
}
\description{
Symbolic interval principal curves limits.
}
\references{
Arce J. and Rodriguez O. (2015) 'Principal Curves and Surfaces
to Interval Valued Variables'. The 5th Workshop on Symbolic
Data Analysis, SDA2015, Orleans, France, November.

Hastie,T. (1984).
Principal Curves and Surface. Ph.D Thesis Stanford University.

Hastie,T. & Weingessel,A. (2014).
princurve - Fits a Principal Curve in Arbitrary Dimension.R package version 1.1--12
http://cran.r-project.org/web/packages/princurve/index.html.

Hastie,T. & Stuetzle, W. (1989). Principal Curves.
Journal of the American Statistical Association, Vol. 84-406, 502--516.

Hastie, T., Tibshirani, R. & Friedman, J. (2008).
The Elements of Statistical Learning; Data Mining, Inference and Prediction. Springer, New York.
}
\seealso{
sym.interval.pc
}
\author{
Jorge Arce.
}
\keyword{Curve}
\keyword{Principal}
