% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.symbolic_df.R
\name{sym.interval.plot}
\alias{sym.interval.plot}
\title{sym.interval.plot}
\usage{
sym.interval.plot(info, col = c("blue"), border = FALSE, show.type = TRUE)
}
\description{
sym.interval.plot
}
\keyword{internal}
