% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.interval.plot.R
\name{sym.interval.plot}
\alias{sym.interval.plot}
\title{sym.interval.plot}
\usage{
sym.interval.plot(info, col = c("blue"), border = FALSE, show.type = TRUE)
}
\description{
<<<<<<< HEAD
sym.interval.plot
=======
Function for plotting one "interval" type value from the symbolic data table
}
\examples{
\dontrun{
data(ex1_db2so)
data.sym <- classic.to.sym(ex1_db2so, concept=c("state", "sex"),
                           col.names = c(county, group, age,age,age,age),
                           variables.types=c(county = type.interval(),
                            group = type.continuous(),
                             age = type.continuous(),
                              age = type.continuous(),
                               age = type.modal(),
                               age = type.histogram()))
sym.interval.plot(data.sym[1,1])
}
}
\author{
Andres Navarro
>>>>>>> master
}
\keyword{internal}
