% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.mean.R
\name{sym.mean}
\alias{sym.mean}
\title{Symbolic Mean}
\usage{
sym.mean(sym.var, method = c('centers', 'interval', 'modal'),
trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{sym.var}{The symbolic variable.}

\item{method}{The method to be use.}

\item{trim}{As in R mean function.}

\item{na.rm}{As in R mean function.}

\item{...}{As in R mean function.}
}
\value{
Return a real number.
}
\description{
This function compute the symbolic mean
}
\examples{
data(example3)
sym.data<-example3
sym.mean(sym.var(sym.data,1))
sym.mean(sym.var(sym.data,2))
sym.mean(sym.var(sym.data,2),method='interval')
sym.mean(sym.var(sym.data,3),method='modal')

}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.

Rodriguez, O. (2000).
Classification et Modeles Lineaires en Analyse des Donnees Symboliques. Ph.D. Thesis,
Paris IX-Dauphine University.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Mean}
\keyword{Symbolic}
