% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.modal.plot.R
\name{sym.modal.plot}
\alias{sym.modal.plot}
\title{Title}
\usage{
sym.modal.plot(info, col = c("blue"), border = FALSE, ylab = TRUE,
  show.type = TRUE, reduce = FALSE)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
