% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.obj.R
\name{sym.obj}
\alias{sym.obj}
\title{Symbolic Object}
\usage{
sym.obj(sym.data, number.sym.obj)
}
\arguments{
\item{sym.data}{Symboli data matrix.}

\item{number.sym.obj}{The number of the row for the symbolic object (case) that we want to get.}
}
\value{
Return a symbolic object with the following internal format: \cr

$M\cr

[1] 5\cr

$var.types\cr

[1] '$C' '$H' '$I' '$H' '$C'\cr

$var.length\cr

[1] 1 5 2 3 1\cr

$var.names\cr

[1] 'F1' 'F2' 'F3' 'F4' 'F5'\cr

$obj.data.vector\cr

F1  M1  M2  M3  M4  M5 F3 F3.1 M1.1 M2.1 M3.1 F5\cr

Case4 -2.1 0.4 0.1 0.1 0.1 0.3  0    2  0.9    0  0.1  0\cr
}
\description{
This function get a symbolic object (row or a case) from a symbolic data table.
}
\examples{
data(example7)
sym.obj(example7,4)
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.

Bock H-H. and Diday E. (eds.) (2000).
Analysis of Symbolic Data. Exploratory methods for extracting statistical information
from complex data. Springer, Germany.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Object}
\keyword{Symbolic}
