% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_radar_plot.R
\name{sym.radar.plot}
\alias{sym.radar.plot}
\title{Internal sym.radar.plot the distence between two rows}
\usage{
sym.radar.plot(
  dat,
  indivs,
  vars,
  rad.main = "",
  rad.legend = "Individuals",
  use.pct = F
)
}
\arguments{
\item{dat}{The symbolic data.}

\item{indivs}{an array that indicates which individuals to use (optional).}

\item{vars}{an array that indicates which variables to use (optional).}

\item{rad.main}{the title of the final plot (optional).}

\item{use.pct}{a logical value indicating if use percentage or real distance for plot.}
}
\value{
A radar plot.
}
\description{
Internal sym.radar.plot the distence between two rows
}
\keyword{For}
\keyword{Interval}
\keyword{Plot}
\keyword{Radar}
\keyword{Symbolic}
\keyword{Variables}
\keyword{internal}
