% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.row.plot.R
\name{sym.row.plot}
\alias{sym.row.plot}
\title{Title}
\usage{
sym.row.plot(data, col = NA, matrix.form = NA, border = FALSE, size = 1,
  title = TRUE, show.type = FALSE, reduce = FALSE)
}
\arguments{
\item{reduce}{}
}
\description{
Title
}
