% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_scatterplot.R
\name{sym.scatterplot}
\alias{sym.scatterplot}
\title{Symbolic Scatter Plot}
\usage{
sym.scatterplot(sym.var.x, sym.var.y, labels = FALSE, ...)
}
\arguments{
\item{sym.var.x}{First symbolic variable}

\item{sym.var.y}{Second symbolic variable.}

\item{labels}{As in R plot function.}

\item{...}{As in R plot function.}
}
\value{
Return a graphics.
}
\description{
This function could be use to plot two symbolic variables in a X-Y plane.
}
\examples{
\dontrun{
data(example3)
sym.data <- example3
sym.scatterplot(sym.data[, 3], sym.data[, 7], col = "blue", main = "Main Title")
sym.scatterplot(sym.data[, 1], sym.data[, 4],
  labels = TRUE, col = "blue",
  main = "Main Title"
)
sym.scatterplot(sym.data[, 2], sym.data[, 6],
  labels = TRUE,
  col = "red", main = "Main Title", lwd = 3
)

data(oils)
sym.scatterplot(oils[, 2], oils[, 3],
  labels = TRUE,
  col = "red", main = "Oils Data"
)
data(lynne1)

sym.scatterplot(lynne1[, 2], lynne1[, 1],
  labels = TRUE,
  col = "red", main = "Lynne Data"
)
}
}
\references{
Bock H-H. and Diday E. (eds.) (2000).
Analysis of Symbolic Data. Exploratory methods for extracting statistical information
from complex data. Springer, Germany.

Rodriguez, O. (2000).
Classification et Modeles Lineaires en Analyse des Donnees Symboliques. Ph.D. Thesis,
Paris IX-Dauphine University.
}
\seealso{
sym.scatterplot3d
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Plot}
\keyword{Symbolic}
