% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.scatterplot3d.R
\name{sym.scatterplot3d}
\alias{sym.scatterplot3d}
\title{Symbolic Scatter Plot 3D}
\usage{
sym.scatterplot3d(sym.var.x, sym.var.y, sym.var.z, labels = FALSE, ...)
}
\arguments{
\item{sym.var.x}{First symbolic variable.}

\item{sym.var.y}{Second symbolic variable.}

\item{sym.var.z}{Third symbolic variable.}

\item{labels}{As in R plot function.}

\item{...}{As in R plot function.}
}
\value{
3D Plot graphic.
}
\description{
This function could be use to plot two symbolic variables in 3D i.e. in a X-Y-Z plane.
}
\examples{
data(lynne1)
class(lynne1) <- c('sym.data.table')
sym.scatterplot3d(lynne1[,1], lynne1[,2], lynne1[,3],
                 color='blue', main='Lynne Data')
}
\references{
Bock H-H. and Diday E. (eds.) (2000).
Analysis of Symbolic Data. Exploratory methods for extracting statistical information
from complex data. Springer, Germany.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{3DPlot}
\keyword{Symbolic}
