% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.symbolic_df.R
\name{sym.set.plot}
\alias{sym.set.plot}
\title{sym.set.plot}
\usage{
sym.set.plot(
  info,
  col = c("blue"),
  border = FALSE,
  show.type = TRUE,
  reduce = FALSE
)
}
\description{
sym.set.plot
}
\keyword{internal}
