% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_interval_umap.R
\name{sym.umap}
\alias{sym.umap}
\alias{sym.umap.symbolic_tbl}
\title{UMAP for Symbolic Data}
\usage{
sym.umap(sym.data, ...)

\method{sym.umap}{symbolic_tbl}(
  sym.data = NULL,
  config = umap::umap.defaults,
  method = c("naive", "umap-learn"),
  preserve.seed = TRUE,
  ...
)
}
\arguments{
\item{sym.data}{symbolic data table}

\item{...}{list of settings; values overwrite defaults from config; see documentation of umap.default for details about available settings}

\item{config}{object of class umap.config}

\item{method}{character, implementation. Available methods are 'naive' (an implementation written in pure R) and 'umap-learn' (requires python package 'umap-learn')}

\item{preserve.seed}{logical, leave TRUE to insulate external code from randomness within the umap algorithms; set FALSE to allow randomness used in umap algorithms to alter the external random-number generator}
}
\description{
This function applies the UMAP algorithm to a symbolic data table.
}
\keyword{Intervals}
