% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_distance-symbolic.set.R,
%   R/symbolic_distance.symbolic_interval.R, R/symbolic_distance.symbolic_tbl.R
\name{symbolic_dist.symbolic_set}
\alias{symbolic_dist.symbolic_set}
\alias{symbolic_dist.symbolic_interval}
\alias{symbolic_dist}
\alias{symbolic_dist.symbolic_tbl}
\title{Distance for symbolic vectors}
\usage{
\method{symbolic_dist}{symbolic_set}(x = sym_set(""), y = sym_set(""))

\method{symbolic_dist}{symbolic_interval}(x = sym_interval(0),
  y = sym_interval(0))

symbolic_dist(x, ...)

\method{symbolic_dist}{symbolic_tbl}(x = NULL, .f = symbolic_dist,
  alpha = 2, w = 1, parallel = F)
}
\arguments{
\item{x}{a symbolic vector or a symbolic_tbl}

\item{y}{a symbolic vector}

\item{.f}{function to compute distances}

\item{alpha}{\itemize{
  \item For alpha = 1, city-block distance
  \item For alpha = 2, euclidean distance
  \item For alpha = ∞ , Chebychev distence
}}

\item{w}{Weight}

\item{parallel}{If False the future strategy ir sequential, otherwise is multiprocess}
}
\description{
Distance for symbolic vectors
}
