% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USCrime}
\alias{USCrime}
\title{Us crime classic data table}
\format{An object of class \code{data.frame} with 1994 rows and 103 columns.}
\source{
http://archive.ics.uci.edu/ml/
}
\usage{
data(USCrime)
}
\description{
Us crime classic data table that can be used to generate symbolic data tables.
}
\examples{
data(USCrime)
us.crime<-USCrime
dim(us.crime)
head(us.crime)
summary(us.crime)
names(us.crime)
nrow(us.crime)
result  <- classic.to.sym(us.crime, concept='state',
                          variables=c('NumInShelters', 'NumImmig'),
                          variables.types=c('$H', '$H'))
result
}
\references{
HASTIE, T., TIBSHIRANI, R. and FRIEDMAN, J. (2008). The Elements of Statistical Learning:
Data Mining, Inference and Prediction. New York: Springer.
}
\keyword{datasets}
