% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VeterinaryData}
\alias{VeterinaryData}
\title{Symbolic interval data example}
\format{
$I Height Height $I Weight Weight \cr

1  $I  120.0  180.0 $I  222.2  354.0\cr

2  $I  158.0  160.0 $I  322.0  355.0\cr

3  $I  175.0  185.0 $I  117.2  152.0\cr

4  $I   37.9   62.9 $I   22.2   35.0\cr

5  $I   25.8   39.6 $I   15.0   36.2\cr

6  $I   22.8   58.6 $I   15.0   51.8\cr

7  $I   22.0   45.0 $I    0.8   11.0\cr

8  $I   18.0   53.0 $I    0.4    2.5\cr

9  $I   40.3   55.8 $I    2.1    4.5\cr

10 $I   38.4   72.4 $I    2.5    6.1\cr
}
\usage{
data(VeterinaryData)
}
\description{
Symbolic data matrix with all the variables of interval type.
}
\examples{
data(VeterinaryData)
VeterinaryData
}
\references{
Billard L. and  Diday E. (2006).
Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}
